Replaced int with float in return from_literal (EdmIntTypTraits)

+++ pyodata/v2/model.py
@@ -458,7 +458,7 @@ class EdmIntTypTraits(TypTraits):
         return int(value)
 
     def from_literal(self, value):
-        return int(value)
+        return float(value)
 
 
 class EdmStructTypTraits(TypTraits):
